feat: 実装クラスをdataモジュールに移動
