+++ modules/data/src/main/java/net/pantasystem/milktea/data/infrastructure/notes/reaction/impl/history/ReactionHistoryCountRecord.kt
@@ -1,0 +1,14 @@
+package net.pantasystem.milktea.data.infrastructure.notes.reaction.impl.history
+
+import androidx.room.ColumnInfo
+import net.pantasystem.milktea.model.notes.reaction.history.ReactionHistoryCount
+
+data class ReactionHistoryCountRecord (
+    val reaction: String,
+    @ColumnInfo(name = "reaction_count")
+    val count: Int
+) {
+    fun toReactionHistoryCount(): ReactionHistoryCount {
+        return ReactionHistoryCount(reaction, count)
+    }
+}
